Make About section compatible with typescript
